backend/pyproject.toml
[tool.poetry]
name = "black8"
version = "0.1.0"
description = ""
authors = ["Dominux <kir.gimranov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "5.0.5"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
